# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API: Clients, Products and Orders",
        "description": "API: Client, Product and Orders",
        "contact": {
            "name": "Fernando Marcelino",
            "url": "https://github.com/fernandoomarcelino",
            "email": "ffernandomarcelino@gmail.com"
        },
        "license": {
            "name": "Private",
            "url": "https://github.com/fernandoomarcelino"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/",
            "description": "L5 Swagger OpenApi dynamic host server"
        }
    ],
    "paths": {
...
        "/api/order": {
            "get": {
                "tags": [
                    "Order"
                ],
                "summary": "Show all",
                "description": "Show all",
                "operationId": "OrderIndex",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Order"
                ],
                "summary": "Create",
                "description": "Create",
                "operationId": "OrderStore",
                "requestBody": {
                    "request": "OrderStoreRequest",
                    "description": "OrderStoreRequest",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OrderStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Missing parameters (validation)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/order/{id}": {
            "get": {
                "tags": [
                    "Order"
                ],
                "summary": "Show detail",
                "description": "Show detail",
                "operationId": "OrderShow",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Order"
                ],
                "summary": "Update",
                "description": "Update",
                "operationId": "OrderUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "request": "OrderUpdateRequest",
                    "description": "OrderUpdateRequest",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OrderUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Missing parameters (validation)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Order"
                ],
                "summary": "Remove",
                "description": "Remove",
                "operationId": "OrderDelete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "OrderStoreRequest": {
                "title": "OrderStoreRequest",
                "description": "The request to create order",
                "properties": {
                    "client_id": {
                        "title": "client_id",
                        "description": "client_id",
                        "type": "integer",
                        "format": "integer"
                    },
                    "products": {
                        "title": "products",
                        "description": "products",
                        "type": "array",
                        "format": "array",
                        "items": {
                            "type": "integer",
                            "items": {}
                        }
                    }
                },
                "type": "object"
            },
            "OrderUpdateRequest": {
                "title": "OrderUpdateRequest",
                "description": "The request to update order",
                "properties": {
                    "client_id": {
                        "title": "client_id",
                        "description": "client_id",
                        "type": "integer",
                        "format": "integer"
                    },
                    "products": {
                        "title": "products",
                        "description": "products",
                        "type": "array",
                        "format": "array",
                        "items": {
                            "type": "integer",
                            "items": {}
                        }
                    }
                },
                "type": "object"
            },
...
            "ValidationErrorResponse": {
                "title": "Validation Error response",
                "description": "The validation error response",
                "properties": {
                    "message": {
                        "title": "message",
                        "description": "error message",
                        "format": "string"
                    },
                    "errors": {
                        "title": "errors",
                        "description": "validation errors in the format: Property name: validation error",
                        "format": "array",
                        "items": {
                            "properties": {
                                "field": {
                                    "title": "field",
                                    "format": "string"
                                },
                                "value": {
                                    "title": "value",
                                    "format": "string"
                                }
                            },
                            "type": "object",
                            "format": "object"
                        }
                    }
                },
                "type": "object"
            }
        }
    },
...
    "security": [
        []
    ]
}